refactor: inspect code 통해 수정 및  패키지 변경
